bin: Add script to automate IPs updates for test environment setup
This commit introduces a script that reads IP and domain pairs from
the output of the 'make hosts' command, and updates
the /etc/hosts, the coredns Corefile and the privnet configs files
accordingly.
The script streamlines the process of setting up both local testing
environments and system tests using Github Actions.

This improves the deployment process, reduces manual intervention
and makes it more reliable and easier to maintain.

Signed-off-by: Oleg Kulachenko <oleg@nspcc.ru>