bin: Add tests for update_hosts.sh script
This commit introduces a tests set for the update_hosts.sh
script.
As per best practices, every part of code, including small utilities, must be
paired by tests. It ensures that the code is resistant to regressions,
which is especially important for code that deals with critical components
of operating system.

The update_hosts.sh script works with the /etc/hosts file, making it all
the more important to have tests for this script. The tests added in this
commit cover various scenarios to make sure that the script behaves as expected
under different conditions.

Signed-off-by: Oleg Kulachenko <oleg@nspcc.ru>