Makefile: configure MaxObjectSize of 1024 for test env
Large objects are defined as this for tests:

  def complex_object_size(max_object_size: int) -> int:
      return max_object_size * int(COMPLEX_OBJECT_CHUNKS_COUNT) + int(COMPLEX_OBJECT_TAIL_SIZE)

max_object_size is taken from the network settings which by default is 64M.
This means that the default large object size is 201327592 or ~200M. And it's
not really needed to test split object functionality, this just wastes some
time and resources for no good reason. 1K is enough here.

Signed-off-by: Roman Khimov <roman@nspcc.ru>